+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added a detail view for workgroups in the admin settings.
 - Enable auto setting of depth in MASL in hydro module.
 - Added functionality to copy extracted text to clipboard in the labeling panel.
+- Added data collection with consent option.
 
 ### Changed